+++ packages/xrpl/test/integration/transactions/singleAssetVault.test.ts
@@ -78,7 +78,7 @@ describe('Single Asset Vault', function () {
         LimitAmount: {
           currency: currencyCode,
           issuer: issuerWallet.classicAddress,
-          value: '1000',
+          value: '9999999999999999e80',
         },
       }
 
@@ -91,7 +91,7 @@ describe('Single Asset Vault', function () {
         Amount: {
           currency: currencyCode,
           issuer: issuerWallet.classicAddress,
-          value: '1000',
+          value: '9999999999',
         },
       }
 
@@ -109,7 +109,7 @@ describe('Single Asset Vault', function () {
           VaultWithdrawalPolicy.vaultStrategyFirstComeFirstServe,
         Data: stringToHex('vault metadata'),
         MPTokenMetadata: stringToHex('share metadata'),
-        AssetsMaximum: '500',
+        AssetsMaximum: '9999900000000000000000000',
       }
 
       await testTransaction(testContext.client, tx, vaultOwnerWallet)
@@ -135,7 +135,7 @@ describe('Single Asset Vault', function () {
         VaultWithdrawalPolicy.vaultStrategyFirstComeFirstServe,
       )
       assert.equal(vault.Data, tx.Data)
-      assert.equal(assetsMaximum, '500')
+      assert.equal(assetsMaximum, '99999e20')
 
       // --- VaultSet Transaction ---
       // Increase the AssetsMaximum to 1000 and update Data