+++ packages/ripple-binary-codec/HISTORY.md
@@ -2,6 +2,9 @@
 
 ## Unreleased
 
+### Fixed
+* Fix STNumber serialization logic to work with large mantissa scale [10^18, 10^19-1].
+
 ## 2.6.0 (2025-12-16)
 
 ### Added